requirements.txt
fastapi
uvicorn
pillow
python-multipart
easyocr
opencv-python-headless
pygments
gradio
requests
openai
transformers
accelerate
torch
black
